Silverpeas Jackrabbit JCA version 2.20.14. Parent POM version 6.4-JR-SNAPSHOT.
For a reason unknown for me, when installing a silverpeas distribution, gradle, used as backbone of the installer, resolves the Silverpeas Jackrabbit JCA rar dependency in the assembly by fetching as transitive dependency the Apache Jackrabbit JCA rar although this dependency isn't declared in the POM of the Silverpeas Jackrabbit JCA rar (only the Apache Jackrabbit JCA jar is declared as dependency)! This occurs even the Jackrabbit JCA jar is declared with a runtime scope.
The consequence is the installer installs some times the wrong Jackrabbit RAR: the Apache one instead of the Silverpeas one.
To avoid this, the Apache Jackrabbit JCA is explicitly excluded from the dependency declaration of the Silverpeas Jackrabbit RAR.